+++ src/Entity/LineItem.php
@@ -39,10 +39,10 @@ class LineItem implements JsonSerializable, EntityInterface
     /** @var string */
     private $slug;
 
-    /** @var DateTimeInterface */
+    /** @var DateTimeInterface|null */
     private $startAt;
 
-    /** @var DateTimeInterface */
+    /** @var DateTimeInterface|null */
     private $endAt;
 
     /** @var int */
@@ -97,7 +97,7 @@ class LineItem implements JsonSerializable, EntityInterface
         return $this->startAt;
     }
 
-    public function setStartAt(DateTimeInterface $startAt): self
+    public function setStartAt(?DateTimeInterface $startAt): self
     {
         $this->startAt = $startAt;
 
@@ -109,7 +109,7 @@ class LineItem implements JsonSerializable, EntityInterface
         return $this->endAt;
     }
 
-    public function setEndAt(DateTimeInterface $endAt): self
+    public function setEndAt(?DateTimeInterface $endAt): self
     {
         $this->endAt = $endAt;