+++ CHANGELOG.md
@@ -7,7 +7,6 @@
 - Added possibility to warm up LTI instance cache separately via `cache-pool` argument of [roster:cache:warmup](docs/cli/cache-warmer-command.md) command.
 - Added possibility to warm up Line Items cache separately.
 - Added `usernames` and `line-item-slugs` input options to [roster:cache:warmup](docs/cli/cache-warmer-command.md) command.
-- Added [roster:modify-entity:line-item:change-state](docs/cli/modify-entity-line-item-change-state-command.md) command.
 - Added [roster:ingest:lti-instance](docs/cli/lti-instance-ingester-command.md) command.
 - Added [roster:ingest:line-item](docs/cli/line-item-ingester-command.md) command.
 - Added [roster:ingest:user](docs/cli/user-ingester-command.md) command.
@@ -15,6 +14,8 @@
 - Added [roster:cache-warmup:lti-instance](docs/cli/lti-instance-cache-warmer-command.md) command.
 - Added [roster:cache-warmup:line-item](docs/cli/line-item-cache-warmer-command.md) command.
 - Added [roster:cache-warmup:user](docs/cli/user-cache-warmer-command.md) command.
+- Added [roster:modify-entity:line-item:change-dates](docs/cli/modify-entity-line-item-change-dates-command.md) command.
+- Added [roster:modify-entity:line-item:change-state](docs/cli/modify-entity-line-item-change-state-command.md) command.
 - Added possibility to use multiple filesystem instances with the help of [Storage registry](docs/storage-registry.md).
 - Added possibility to launch assignments with [LTI 1.3](http://www.imsglobal.org/spec/lti/v1p3/).
 - Added possibility to process a `basic outcome replaceResult` request using LTI 1.3 flow.